split the C interface into a separate file
